Fixed bug with adding a job app
When adding a job app, the form would try to submit after you filled out any of the fields. Since there are 2 required fields, it was getting an error and trying to redirect to "new.html.erb", which doesn't exist anymore, and even if it did, it's still very annoying. I changed the form for a new job_app have a submit button instead of automatically submitting, so it doesn't redirect